+++ crates/deadpool-runtime/CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Add support for `smol` version 2.0
+- Mark `rt_async-std_1` feature as deprecated
+
 ## [0.1.4] - 2024-05-24
 
 - Add `LICENSE-APACHE` and `LICENSE-MIT` files to published crates